requirements.txt
feedparser
httplib2
google-genai
python-dateutil